Separate MacOs Pkg tool into cli and library projects
