+++ src/Microsoft.DotNet.MacOsPkg/Core/ExecuteHelper.cs
@@ -5,7 +5,7 @@ using System;
 using System.Diagnostics;
 using System.IO;
 
-namespace Microsoft.DotNet.MacOsPkg
+namespace Microsoft.DotNet.MacOsPkg.Core
 {
     public static class ExecuteHelper
     {
